added cast to signed/unsigned comparison

+++ include/sf2/formats/json_reader.hpp
@@ -220,7 +220,7 @@ namespace format {
 		if constexpr(std::numeric_limits<T>::is_signed) {
 			if(negativ)
 				return static_cast<T>( -static_cast<T>(val) );
-			else if(val<=std::numeric_limits<T>::max())
+			else if(val<=static_cast<std::uint64_t>(std::numeric_limits<T>::max()))
 				return static_cast<T>(val);
 
 		} else {
